% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clumped_data}
\alias{clumped_data}
\title{Example clumped data}
\format{
A data frame with 387 rows and 16 variables.
}
\usage{
clumped_data
}
\description{
Harmonized and clumped GWAS summary statistics for LDLC and MI.
}
\keyword{datasets}
