% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XMR.R
\name{fit_XMR}
\alias{fit_XMR}
\title{A function for implementing XMR.}
\usage{
fit_XMR(
  data,
  exposure = "exposure",
  outcome = "outcome",
  SigmaX = NULL,
  tau.sq = NULL,
  pi0 = NULL,
  C = diag(3),
  Omega0 = matrix(0, 3, 3),
  tol1 = 1e-06,
  tol2 = 1e-06,
  Threshold = 5e-05,
  ELBO = F,
  verbose = T,
  min_thres = 1e-04,
  max_iter = 5000
)
}
\arguments{
\item{data}{data.frame that at least contains the following columns:
b.exp.pop1, b.exp.pop2, b.out.pop2, se.exp.pop1, se.exp.pop2, se.out.pop2,
L2.pop1, L12, L2.pop2. b: SNP-trait effect; se: the standard error of b
(exp: exposure, out: outcome, pop1: the auxiliary population,
pop2: the target population); L2.pop1: LD score in the auxiliary population;
L2.pop2: LD score in the target population; L12: cross-population LD score
between pop1 and pop2.}

\item{exposure}{exposure name.}

\item{outcome}{outcome name.}

\item{SigmaX}{initial value for SigmaX, default `NULL` will use the default
initialize procedure.}

\item{tau.sq}{initial value for tau.sq, default `NULL` will use the default
initialize procedure.}

\item{pi0}{initial value for pi0, default `NULL` will use the default
initialize procedure.}

\item{C}{the estimated C matrix capturing the effects of sample structure.
Default `diag(3)`.}

\item{Omega0}{the estimated variance-covariance matrix of polygenic effects.
Default `matrix(0, 3, 3)`.}

\item{tol1}{tolerance of the 1st fitting round fixing beta=0, default `1e-06`.}

\item{tol2}{tolerance of the 2nd fitting round which fits beta, default `1e-06`.}

\item{Threshold}{modified IV selection threshold for correction of selection bias.}

\item{ELBO}{logical, whether to track the evidence lower bound or not.
If `FALSE`, check the maximum likelihood instead. Default `FALSE`.}

\item{verbose}{logical, whether to print results. Default `TRUE`.}

\item{min_thres}{lower bound of pi0 to avoid log(0); pi0 smaller than this
value during fitting will be truncated to it. Default `1e-4`.}

\item{max_iter}{maximum iteration rounds of the variational-EM algorithm.
Default `5000`.}
}
\value{
A list with the following elements:
\describe{
  \item{exposure}{exposure of interest}
  \item{outcome}{outcome of interest}
  \item{beta}{causal effect estimate}
  \item{beta.se}{standard error}
  \item{beta.pvalue}{p-value}
  \item{tau.sq}{variance of the outcome effect in the causal inference module}
  \item{SigmaX}{covariance matrix of exposure effects in the causal inference module}
  \item{post}{posterior probability of the validity of each input IV}
  \item{nIV}{number of input IVs}
  \item{nvalid}{number of valid IVs detected by XMR}
  \item{pi0}{the probability of an SNP with foreground signals after selection}
  \item{Threshold}{input modified IV selection threshold for correction of selection bias}
  \item{fit1_elbos}{records of the ELBO changes in the 1st fitting round}
  \item{fit2_elbos}{records of the ELBO changes in the 2nd fitting round}
  \item{fit1_likelis}{records of the likelihood changes in the 1st fitting round}
  \item{fit2_likelis}{records of the likelihood changes in the 2nd fitting round}
}
}
\description{
XMR: a cross-population Mendelian randomization method for
  causal inference using genome-wide summary statistics.
  XMR uses a variational EM algorithm for the estimation of parameters.
  XMR uses likelihood ratio test for inference.
}
\examples{
exposure <- "LDLC"
outcome  <- "MI"
threshold <- 5e-05
N1 <- 343621 # EUR sample size
N2 <- 72866  # EAS sample size
t0 <- abs(qnorm(threshold / 2))
dt <- 0.13 / (sqrt(N2 / N1))
modified_threshold <- 2 * (1 - pnorm(abs(t0 + dt))) # modified threshold
data(C)
data(Omega)
data(clumped_data)
XMR_res <- fit_XMR(
  data = clumped_data,
  C = C,
  Omega0 = Omega,
  Threshold = modified_threshold,
  tol1 = 1e-07,
  tol2 = 1e-07,
  min_thres = 1e-2
)
}
